+++ apps/throttle/src/views/ProgrammingView.vue
@@ -1,0 +1,9 @@
+<script setup lang="ts">
+  import CvProgrammer from '@/programming/CvProgrammer.vue'
+</script>
+
+<template>
+  <main class="@container min-h-screen forest-background bg-opacity-50 bg-fixed overflow-auto p-2 md:p-4">
+    <CvProgrammer />
+  </main>
+</template>