+++ packages/modules/index.ts
@@ -43,3 +43,7 @@ export { useBlocks } from './sensors/useBlocks'
 export { useAutomations } from './sensors/useAutomations'
 export * from './sensors/types'
 export * from './sensors/constants'
+
+// CV module
+export * from './cv/types'
+export * from './cv/constants'